Move `buildLineRange` to `StringUtil`
